Improve TypeID interfaces (#195)
## Summary
Improvements to TypeID interfaces:
- Add all the standard TypeID methods to the `Subtype` interfaces. When
changing code in axiom, noticed that would be useful.
- Make `SubtypePtr` public. Occasionally useful, if one needs to create
a 'constructor'-like function outside the `typeid` package.
- Rename `Type()` to `Prefix()`. Been wanting to do this change before.
After using typeids, I think `Prefix()` feels more natural and it pairs
nicely with the existing `Suffix()`. This change is *not* backwards
compatible, but neither was the generics change we did. I'm hoping I can
get this in now, and then tag the next release as `v1`.

## How was it tested?
Ran all tests.