Upgrade go, devbox-install action, update README (#437)
## Community Contribution License
All community contributions in this pull request are licensed to the
project maintainers
under the terms of the [Apache 2
license](https://www.apache.org/licenses/LICENSE-2.0).
By creating this pull request I represent that I have the right to
license the
contributions to the project maintainers under the Apache 2 license.

## Summary
Remove the "open in devbox" link from the TySON readme. I think you had
remove this previously by directly doing it in the TySON repo, but I had
to revert it to allow publishing from the opensource repo to work again.
Re-doing the change via the opensource monorepo with this PR.

Because CICD seems to be failing, also upgrading the go version.

## How was it tested?
N/A